chore(colors.py): move colors.py to `src/simulat/core/`
