+++ src/simulat/core/colors.py
